delete useless props
